# Optimize graph MLE tests