* Reduces number of bootstrap iterations: 100 -> 10
* Avoids generating instance for digraph assertion test
* Sets RNG seed in remaining tests using randomness
* Fixes typo leading to trivial assertion

Before:
```
148.30s call     tests/test_mle.py::test_infer_node_dgs_w_error_invariant_wrt_edge_order
65.64s call     tests/test_mle.py::test_infer_node_dgs_w_error
21.63s setup    tests/test_mle.py::test_infer_node_vals_and_errs_networkx_requires_di_graph[0]
21.40s call     tests/test_mle.py::test_infer_node_vals_and_errs_networkx[0]
```

After:
```
16.11s call     tests/test_mle.py::test_infer_node_dgs_w_error_invariant_wrt_edge_order
7.08s call     tests/test_mle.py::test_infer_node_dgs_w_error
5.10s call     tests/test_mle.py::test_infer_node_vals_and_errs_networkx[0]
```